consolidate list tests
